Try to gather enough information to dynamically retrieve the run logs
Starting point is roughly:
  https://api.github.com/repos/check-spelling/zio-json/actions/runs/3102202564

One can reach:
  https://api.github.com/repos/check-spelling/zio-json/actions/runs/3102202564/jobs

With some effort, one can tease out the second job and get some fields:

  name: "Check Spelling"

The active step in steps:[]

We can try to cheat by assuming that it matches our action.yml, or w/ some effort we can determine it...

In it, we want:
The step number.

This will let us assemble:

  Check Spelling/2_check-spelling.txt

Which we can use to extract from the zip file that corresponds to:
  https://api.github.com/repos/check-spelling/zio-json/actions/runs/3102202564/logs

Which can be retrieved using `gh api`.