Adjust minified-file handling
Treat runs of 3 or more `=` characters as if zero length strings.

Currently, `=` is the "replacement" character used by patterns/reject.
This does mean that a very wide underscore line would be ignored, but
generally that line would be paired with a line preceding it which
itself should trigger the complaint.