Sarif: handle encoding
Sarif generation is a bit of a mess...

Each result is currently constructed as a string `$result_json`.
That string is then converted into a json happy object with decode_json.
Eventually all of those json objects are serialized back into json.

There is a pass in parse_warnings that tries to json escape things,
but without an `encode_json`, its encoding would be immediately
undone by that `decode_json`.

A more correct fix would be to properly construct json objects.