Replace Microsoft Outlook hack
Microsoft Outlook.app doesn't understand:
<details><summary></summary></details>

... specifically, it treats them as a span which
results in some truly lousy rendering.

The previous hack of introducing an emoji at least
prevented users from seeing a running sentence.

The new hack is to introduce an empty `<p></p>`.
Web browsers ignore it (because they already have
a break), but Microsoft Outlook.app treats it as a
paragraph break or something which is desirable.